Proof of Concept
This is a basic working version of the tool that I've wanted to have for the last couple of years: have a student drop a Git commit on to it and it renders out the relevant source code (not just the code changed in the commit, but all the code in the project affected by the commit). Hopefully this gets me closer to being able to use [SpeedGrader](https://community.canvaslms.com/docs/DOC-2583) in Canvas to mark up student code that has been committed to GitHub!